feat(export): performance improvements (#229)
Export seems extremely inefficient as it is currently.

Tested with P. dataset from 01JAN2024 which contains 361 entries (from admin account) on dev.
It currently takes around 5 seconds.
Test on prod with 1k+ seems to be far worse.
So it might explain user feedback that export is "not working".

The goal of this PR is to bypass usual calls and craft an optimized query especially for export by removing multiple calls and do a big huge list of joins.
It's ugly, but seems promising and should scale better.

Code was done a while back but needs more testing before merging it.

Reviewed-on: https://git.tristess.app/alexandresoro/ouca/pulls/229